project_tracer/backend/requirements.txt
fastapi
uvicorn[standard]
motor
httpx
beautifulsoup4
pydantic
python-dotenv
langchain
langchain-google-genai
requests
